Fixed login feature bugs and added image link to clothingItem model

+++ server/controllers/clothingItem.controller.js
@@ -1,11 +1,11 @@
-import ClothingItem from "../models/clothingItem.model.js";
+import ClothingItem from "../models/clothingItem.model.js"
 
 // Create
 export const createClothingItem = async (req, res) => {
     try {
         const NEW_CLOTHINGITEM = await ClothingItem.create({
-        ...req.body,
-        user: req.user.id,
+            ...req.body,
+            user: req.user.id
         })
         res.status(201).json(NEW_CLOTHINGITEM)
     } catch (error) {res.status(400).json(error)}
@@ -14,7 +14,9 @@ export const createClothingItem = async (req, res) => {
 // Read
 export const getAllClothingItems = async (req, res) => {
     try {
+        console.log(req.user)
         const CLOTHINGITEMS = await ClothingItem.find({ user: req.user.id })
+        if (!CLOTHINGITEMS) return res.status(404).json({ error: "Clothing items not found" })
         res.status(201).json(CLOTHINGITEMS)
     } catch (error) {res.status(400).json(error)}
 }
@@ -22,6 +24,7 @@ export const getAllClothingItems = async (req, res) => {
 export const getClothingItemById = async (req, res) => {
     try {
         const CLOTHINGITEM = await ClothingItem.findById(req.params.id)
+        if (!CLOTHINGITEM) return res.status(404).json({ error: "Clothing item not found" })
         if (CLOTHINGITEM.user.toString() !== req.user.id) {
             return res.status(404).json({ error: "Access Denied" });
         }
@@ -31,15 +34,17 @@ export const getClothingItemById = async (req, res) => {
 
 // Update
 export const updateClothingItemById = async (req, res) => {
-    const options = {
-        new: true,
-        runValidators: true
-    }
     try {
-        const UPDATED_CLOTHINGITEM = await ClothingItem.findByIdAndUpdate(req.params.id, req.body, options)
-        if (UPDATED_CLOTHINGITEM.user.toString() !== req.user.id) {
-            return res.status(404).json({ error: "Access Denied" });
+        const CLOTHINGITEM = await ClothingItem.findById(req.params.id);
+        if (!CLOTHINGITEM) return res.status(404).json({ error: "Clothing item not found" })
+        if (CLOTHINGITEM.user.toString() !== req.user.id) {
+            return res.status(404).json({ error: "Access Denied" })
         }
+        const options = {
+            new: true,
+            runValidators: true
+        }
+        const UPDATED_CLOTHINGITEM = await ClothingItem.findByIdAndUpdate(req.params.id, req.body, options)
         res.status(201).json(UPDATED_CLOTHINGITEM)
     } catch (error) {res.status(400).json(error)}
 }
@@ -47,10 +52,12 @@ export const updateClothingItemById = async (req, res) => {
 // Delete
 export const deleteClothingItemById = async (req, res) => {
     try {
-        const DELETED_CLOTHINGITEM = await ClothingItem.findByIdAndDelete(req.params.id)
-        if (DELETED_CLOTHINGITEM.user.toString() !== req.user.id) {
-            return res.status(404).json({ error: "Access Denied" });
+        const DELETED_CLOTHINGITEM = await ClothingItem.findById(req.params.id)
+        if (!DELETED_CLOTHINGITEM) return res.status(404).json({ error: "Clothing item not found" })
+        if (item.user.toString() !== req.user.id) {
+            return res.status(403).json({ error: "Access Denied" })
         }
+        await ClothingItem.findByIdAndDelete(req.params.id)
         res.status(201).json(DELETED_CLOTHINGITEM)
     } catch (error) {res.status(400).json(error)}
 }
